Move contributions grid to contributions module
